requirements.txt
opencv-python
numpy
pyautogui
mss
keyboard
pillow
playsound
